+++ entity-view/api/src/main/java/com/blazebit/persistence/view/MappingCorrelatedSimple.java
@@ -67,6 +67,15 @@ public @interface MappingCorrelatedSimple {
     String correlationResult() default "";
 
     /**
+     * The associations of the entity that should be fetched.
+     * This is only valid if the mapping refers to an entity and is mapped as attribute with the original type.
+     *
+     * @return The JPQL expression mapping
+     * @since 1.2.0
+     */
+    String[] fetches() default {};
+
+    /**
      * The fetch strategy to use for correlation.
      *
      * @return The correlation fetch strategy