+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/basic/model/FetchesDocumentView2.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2014 - 2017 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.view.testsuite.basic.model;
+
+import com.blazebit.persistence.view.EntityView;
+import com.blazebit.persistence.view.Mapping;
+import com.blazebit.persistence.view.testsuite.entity.Document;
+import com.blazebit.persistence.view.testsuite.entity.Person;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * @author Christian Beikov
+ * @since 1.2.0
+ */
+@EntityView(Document.class)
+public interface FetchesDocumentView2 extends IdHolderView<Long> {
+
+    @Mapping(fetches = "ownedDocuments")
+    Map<Integer, Person> getContacts();
+
+    @Mapping(fetches = "ownedDocuments")
+    List<Person> getPersonList();
+}